support arena and add Allocator api
