+++ structtype.go
@@ -53,19 +53,19 @@ func init() {
 	SetCustomFunc(reflect.TypeOf(sync.Mutex{}), emptyCloneFunc)
 	SetCustomFunc(reflect.TypeOf(sync.RWMutex{}), emptyCloneFunc)
 	SetCustomFunc(reflect.TypeOf(sync.WaitGroup{}), emptyCloneFunc)
-	SetCustomFunc(reflect.TypeOf(sync.Cond{}), func(old, new reflect.Value) {
+	SetCustomFunc(reflect.TypeOf(sync.Cond{}), func(allocator *Allocator, old, new reflect.Value) {
 		// Copy the New func from old value.
 		oldL := old.FieldByName("L")
-		newL := noState.clone(oldL)
+		newL := allocator.Clone(oldL)
 		new.FieldByName("L").Set(newL)
 	})
-	SetCustomFunc(reflect.TypeOf(sync.Pool{}), func(old, new reflect.Value) {
+	SetCustomFunc(reflect.TypeOf(sync.Pool{}), func(allocator *Allocator, old, new reflect.Value) {
 		// Copy the New func from old value.
 		oldFn := old.FieldByName("New")
-		newFn := noState.clone(oldFn)
+		newFn := allocator.Clone(oldFn)
 		new.FieldByName("New").Set(newFn)
 	})
-	SetCustomFunc(reflect.TypeOf(sync.Map{}), func(old, new reflect.Value) {
+	SetCustomFunc(reflect.TypeOf(sync.Map{}), func(allocator *Allocator, old, new reflect.Value) {
 		if !old.CanAddr() {
 			return
 		}
@@ -74,13 +74,13 @@ func init() {
 		oldMap := old.Addr().Interface().(*sync.Map)
 		newMap := new.Addr().Interface().(*sync.Map)
 		oldMap.Range(func(key, value interface{}) bool {
-			k := Clone(key)
-			v := Clone(value)
-			newMap.Store(k, v)
+			kVal := allocator.Clone(reflect.ValueOf(key))
+			vVal := allocator.Clone(reflect.ValueOf(value))
+			newMap.Store(kVal.Interface(), vVal.Interface())
 			return true
 		})
 	})
-	SetCustomFunc(reflect.TypeOf(atomic.Value{}), func(old, new reflect.Value) {
+	SetCustomFunc(reflect.TypeOf(atomic.Value{}), func(allocator *Allocator, old, new reflect.Value) {
 		if !old.CanAddr() {
 			return
 		}
@@ -88,8 +88,9 @@ func init() {
 		// Clone value inside atomic.Value.
 		oldValue := old.Addr().Interface().(*atomic.Value)
 		newValue := new.Addr().Interface().(*atomic.Value)
-		v := Clone(oldValue.Load())
-		newValue.Store(v)
+		v := oldValue.Load()
+		val := allocator.Clone(reflect.ValueOf(v))
+		newValue.Store(val.Interface())
 	})
 }
 
@@ -100,8 +101,8 @@ func init() {
 // If a struct type contains scalar type fields only, the struct will be marked as scalar automatically.
 //
 // Here is a list of types marked as scalar by default:
-//     * time.Time
-//     * reflect.Value
+//   - time.Time
+//   - reflect.Value
 func MarkAsScalar(t reflect.Type) {
 	for t.Kind() == reflect.Ptr {
 		t = t.Elem()
@@ -118,8 +119,8 @@ func MarkAsScalar(t reflect.Type) {
 // If t is not a pointer, MarkAsOpaquePointer ignores t.
 //
 // Here is a list of types marked as opaque pointers by default:
-//     * `elliptic.Curve`, which is `*elliptic.CurveParam` or `elliptic.p256Curve`;
-//     * `reflect.Type`, which is `*reflect.rtype` defined in `runtime`.
+//   - `elliptic.Curve`, which is `*elliptic.CurveParam` or `elliptic.p256Curve`;
+//   - `reflect.Type`, which is `*reflect.rtype` defined in `runtime`.
 func MarkAsOpaquePointer(t reflect.Type) {
 	if t.Kind() != reflect.Ptr {
 		return
@@ -133,11 +134,11 @@ func MarkAsOpaquePointer(t reflect.Type) {
 // which `new.CanSet()` and `new.CanAddr()` is guaranteed to be true.
 //
 // Func must update the new to return result.
-type Func func(old, new reflect.Value)
+type Func func(allocator *Allocator, old, new reflect.Value)
 
 // emptyCloneFunc is used to disable shadow copy.
 // It's useful when cloning sync.Mutex as cloned value must be a zero value.
-func emptyCloneFunc(old, new reflect.Value) {}
+func emptyCloneFunc(allocator *Allocator, old, new reflect.Value) {}
 
 // SetCustomFunc sets a custom clone function for type t.
 // If t is not struct or pointer to struct, SetCustomFunc ignores t.
@@ -166,7 +167,7 @@ type structFieldType struct {
 }
 
 // NewFrom creates a new value of src.Type() and shadow copies all content from src.
-func (st *structType) Copy(src, nv reflect.Value) {
+func (st *structType) Copy(allocator *Allocator, src, nv reflect.Value) {
 	dst := nv.Elem()
 
 	if st.fn != nil {
@@ -174,7 +175,7 @@ func (st *structType) Copy(src, nv reflect.Value) {
 			src = forceClearROFlag(src)
 		}
 
-		st.fn(src, dst)
+		st.fn(allocator, src, dst)
 		return
 	}
 
@@ -251,10 +252,15 @@ func isScalar(k reflect.Kind) bool {
 		reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64, reflect.Uintptr,
 		reflect.Float32, reflect.Float64,
 		reflect.Complex64, reflect.Complex128,
-		reflect.String, reflect.Func,
+		reflect.Func,
 		reflect.UnsafePointer,
 		reflect.Invalid:
 		return true
+
+	case reflect.String:
+		// If arena is not enabled, string can be copied as scalar safely
+		// as it's immutable by design.
+		return !arenaIsEnabled
 	}
 
 	return false
