+++ structtype_sample_test.go
@@ -55,21 +55,24 @@ func ExampleSetCustomFunc() {
 	}
 
 	// Filter nil values in Data when cloning old value.
-	SetCustomFunc(reflect.TypeOf(MyStruct{}), func(old, new reflect.Value) {
+	SetCustomFunc(reflect.TypeOf(MyStruct{}), func(allocator *Allocator, old, new reflect.Value) {
 		// The new is a zero value of MyStruct.
 		// We can get its address to update it.
 		value := new.Addr().Interface().(*MyStruct)
 
 		// The old is guaranteed to be a MyStruct.
 		// As old.CanAddr() may be false, we'd better to read Data field directly.
-		data := old.FieldByName("Data").Interface().([]interface{})
+		data := old.FieldByName("Data")
+		l := data.Len()
 
-		for _, v := range data {
-			if v == nil {
+		for i := 0; i < l; i++ {
+			val := data.Index(i)
+
+			if val.IsNil() {
 				continue
 			}
 
-			n := Clone(v)
+			n := allocator.Clone(val).Interface()
 			value.Data = append(value.Data, n)
 		}
 	})
